move request, response /src/data to /controllers/definitions/
