+++ src/controllers/Series.ts
@@ -5,7 +5,7 @@ import { logger, LoggerHelper } from '@modules/winston';
 import IngredientService from '@services/IngredientService';
 import SeriesService from '@services/SeriesService';
 
-import { PagingRequestDTO } from '@request/common';
+import { PagingRequestDTO } from '@src/controllers/definitions/request/common';
 
 import { ResponseDTO } from '@response/common';
 import { SeriesResponse, SeriesFilterResponse } from '@response/series';
@@ -146,7 +146,7 @@ const getIngredients: RequestHandler = (
  *       produces:
  *       - application/json
  *       parameters:
- *       - name: itemSize
+ *       - name: requestSize
  *         in: query
  *         type: integer
  *         required: false
