+++ src/controllers/User.ts
@@ -24,7 +24,10 @@ import StatusCode from '@utils/statusCode';
 
 import UserService from '@services/UserService';
 
-import { UserRegisterRequest, UserEditRequest } from '@request/user';
+import {
+    UserRegisterRequest,
+    UserEditRequest,
+} from '@src/controllers/definitions/request/user';
 
 import { ResponseDTO, SimpleResponseDTO } from '@response/common';
