+++ src/controllers/definitions/response/series.ts
