+++ src/modules/swagger.ts
@@ -60,8 +60,8 @@ const options: swaggerJSDoc.Options = {
         './src/controllers/*.ts',
         './src/controllers/*.js',
         './src/middleware/auth.ts',
-        './src/data/request/*.ts',
-        './src/data/response/*.ts',
+        './src/controller/definitions/request/*.ts',
+        './src/controller/definitions/response/*.ts',
         './api/swagger.yaml',
         './api/definitions/*.yaml',
     ],
