+++ tsconfig.json
@@ -70,8 +70,8 @@
             "@modules/*" : ["modules/*"],
             "@utils/*" : ["utils/*"],
             "@libs/*" : ["lib/*"],
-            "@request/*" : ["data/request/*"],
-            "@response/*" : ["data/response/*"],
+            "@request/*" : ["controllers/definitions/request/*"],
+            "@response/*" : ["controllers/definitions/response/*"],
             "@middleware/*" : ["middleware/*"],
             "@configs/*" : ["config/*"]
         },                                 /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
